requirements.txt
PyPDF2
beautifulsoup4
python-docx
requests
json
